Refactor FastAPI app into modular packages
