+++ app/models/common.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel, EmailStr
+
+class PlacementRequest(BaseModel):
+    student_email: EmailStr
+    job_code: str
+
