+++ app/models/metrics.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class StudentLoadTimeMetric(BaseModel):
+    role: str
+    duration: float
+
